requirements.txt
-i https://pypi.org/simple
certifi==2019.3.9
chardet==3.0.4
idna==2.8
oauthlib==3.0.1
pysocks==1.6.8
python-dotenv==0.10.1
requests-oauthlib==1.2.0
requests==2.21.0
six==1.12.0
tweepy==3.7.0
twython==3.7.0
urllib3==1.24.2
